requirements.txt
requests
discord.py
instaloader
ffmpeg-python
python-dotenv
